Separate Datasets/Datafiles Inputs tabs at top level
In preference to having an extra level of tabs under a single Inputs
tab, use separate tabs at the top level.
There is still only one Details subpanel. It appears when a dataset or
datafile is clicked, and remains in view "under" all three tabs. This
makes it easier to copy/paste/drag items from the details panel into the
options.