Report Cancel failures back to user
Similar to the changes to Submit: change the service to pass Cancel
error response messages back, and change my-jobs.controller to display
them to the user. The most likely failure is when the job has actually
Completed but the IJP has yet to detect this, and still allows the user
to Cancel; the display advises the user to check again in "a couple of
minutes".