Display count of active jobs
Update the My Jobs panel heading to show the number of 'active' jobs
(jobs that are queued or executing).
Motivation: the user should not log out of topcat while there are still
active jobs (that may require the sessionId to be valid). We cannot
prevent this (yet), so a reminder will have to suffice.